+++ app/commands/deploy.js
@@ -1,6 +1,6 @@
 import { COMMAND_DEPLOY } from '../../constants/command.js';
 import { deploy } from '../../services/vercel.js';
-import Event from '../models/event.js';
+import Event from '../event.js';
 
 const isDeployCommand = (event) => event.isCommand(COMMAND_DEPLOY);
 
@@ -13,9 +13,8 @@ const execDeployCommand = async (event) => {
     await deploy();
     event.sendText('deploying');
   } catch (err) {
-    event
-      .sendText(err.message)
-      .sendText(err.response.data.error.message);
+    event.sendText(err.message);
+    if (err.response) event.sendText(err.response.data.error.message);
   }
   return event;
 };