+++ utils/generate-completion.js
@@ -1,10 +1,21 @@
 import config from '../config/index.js';
-import { createCompletion, FINISH_REASON_STOP, PARTICIPANT_AI } from '../services/openai.js';
+import { createCompletion, FINISH_REASON_STOP } from '../services/openai.js';
 
 class Completion {
-  constructor({ prompt, text }) {
-    this.prompt = prompt;
+  text;
+
+  finishReason;
+
+  constructor({
+    text,
+    finishReason,
+  }) {
     this.text = text;
+    this.finishReason = finishReason;
+  }
+
+  get isFinishReasonStop() {
+    return this.finishReason === FINISH_REASON_STOP;
   }
 }
 
@@ -16,15 +27,14 @@ class Completion {
  */
 const generateCompletion = async ({
   prompt,
-  text = '',
 }) => {
-  const completion = new Completion({ prompt, text });
-  if (config.APP_ENV !== 'production') return completion;
+  if (config.APP_ENV !== 'production') return new Completion({ text: 'OK' });
   const { data } = await createCompletion({ prompt });
   const [choice] = data.choices;
-  completion.prompt += choice.text.trim();
-  completion.text += choice.text.replace(PARTICIPANT_AI, '').replace(':', '').replace('：', '').trim();
-  return choice.finish_reason === FINISH_REASON_STOP ? completion : generateCompletion(completion);
+  return new Completion({
+    text: choice.text.trim(),
+    finishReason: choice.finish_reason,
+  });
 };
 
 export default generateCompletion;